[add]内蔵クラスを別ファイル化
